requirements.txt
# Core dependencies
numpy>=1.20.0
Pillow>=8.0.0
cryptography>=36.0.0

# For testing
pytest>=6.0.0

# For Jupyter notebooks
jupyter>=1.0.0
matplotlib>=3.3.0
ipywidgets>=7.6.0
